encrypt password

+++ controller/dto/request/user_request.go
@@ -4,5 +4,5 @@ type UserRequest struct {
 	Email    string `json:"email" binding:"required,email"`
 	Password string `json:"password" binding:"required,min=6,containsany=!@#$%*"`
 	Name     string `json:"name" binding:"required,min=2,max=255"`
-	Age      int    `json:"age" binding:"required,numeric,min=1,max=120"`
+	Age      int8    `json:"age" binding:"required,numeric,min=1,max=120"`
 }
